Update domain models, fix tests, update agency loaders
